% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drrglm.R
\name{drr}
\alias{drr}
\alias{drrglm}
\title{Doubly Regularized Matrix-Variate GLMs}
\usage{
drrglm(
  x,
  y,
  family,
  lambda1,
  lambda2,
  C0 = NULL,
  tol = 0.001,
  maxIter = 300,
  verbose = FALSE
)
}
\arguments{
\item{x}{\eqn{p_1 \times p_2 \times N} numeric array with \strong{mean zero}.}

\item{y}{Numeric vector of length \eqn{N}.}

\item{family}{See \link{glm} and \link{family}.}

\item{lambda1}{\eqn{\lambda_1}.}

\item{lambda2}{\eqn{\lambda_2}.}

\item{C0}{Initialization to \eqn{C}.}

\item{tol}{Convergence tolerance.}

\item{maxIter}{Maximal step of iterations.}

\item{verbose}{Print iterations?}
}
\value{
\code{list(L, S, Lsv, iter, lambda1, lambda2, isConvergent)}.
}
\description{
Solve the following problem
\deqn{
  \min_{L, S} \Big\lbrace \frac{1}{N} 
     \sum_{n=1}^N \ell(y_n, X_n) + \lambda_1 \|L\|_* + \lambda_2 \|S\|_1 
  \Big\rbrace,
}
where \eqn{\ell(y_n, X_n)} refers to the negative-log-likelihood on
\eqn{(y_n, X_n)} under pre-specified GLM.

In linear model, the problem has the form
\deqn{
     \min_{L, S} \Big\lbrace \frac{1}{2N} \sum_{n=1}^N 
         \big( y_n - \text{tr}(X_n'C) \big)^2 + 
         \lambda_1 \|L\|_* + \lambda_2 \|S\|_1 \Big\rbrace,
     ~~\text{s.t.}~~ C=L+S.
}
}
\examples{
set.seed(2025)
r0 <- 13
c0 <- 17
N <- 500
x <- array(runif(r0*c0*N), c(r0, c0, N))
y <- rnorm(N)

family <- "gaussian"
lambda1 <- 0.15
lambda2 <- 0.04

system.time( egg <- drrglm(x, y, family, lambda1, lambda2, verbose=TRUE) )

}
