% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEG.R
\docType{data}
\name{EEG}
\alias{EEG}
\title{ElectroEncephaloGraphy Data}
\format{
\code{list(alcholic, control)}.
}
\usage{
EEG
}
\description{
This data arises from a large study to examine EEG
correlates of genetic predisposition to alcoholism.
See details.
}
\details{
The original data are fully open-access and available in UCI Machine
Learning Repository (\url{http://kdd.ics.uci.edu/databases/eeg/}).
It includes two groups of subjects: 77 alcoholic and 45 control.
In the original study, each subject was exposed to either a single
stimulus (S1) or two stimuli (S1 and S2), which were pictures chosen from
the 1980 Snodgrass and Vanderwart picture set.
Each subject underwent 120 trials under each condition.

Here we provide this preprocessed data of the averages of 120 trials under
S1 condition, which has been studied in several literature.
The dataset is structured as a list containing two arrays,
\itemize{
\item \code{EEG$alcholic}: Array of dimensions \verb{256 x 64 x 77}.
\item \code{EEG$control}:  Array of dimensions \verb{256 x 64 x 45}.
}
}
\examples{
data(EEG, package="drrglm")

}
\references{
If you use this dataset, we would be very grateful if you could cite both
the original data source and our work:

Zengchao Xu, Shan Luo, and Binyan Jiang. "Doubly Regularized Matrix-Variate Regression". \emph{Submitted}.
}
\keyword{datasets}
