% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-ratio-rank.R
\name{get_gr_rank}
\alias{get_gr_rank}
\title{Identify Rank using Growth-Ratio Criterion}
\usage{
get_gr_rank(sv, tol = 1e-04)
}
\arguments{
\item{sv}{Vector of singular values.}

\item{tol}{The singular values smaller than tol is treated as zero.}
}
\value{
An integer of rank.
}
\description{
\loadmathjax
Let \mjseqn{V(k) = \sum_{j=k+1}^r \sigma_j (\hat{L})},
\mjsdeqn{GR(k) = \frac{
     \log(1 + \sigma_k(\hat{L}) / V(k))}{
     \log(1 + \sigma_{k+1}(\hat{L})/V(k+1))}
}
}
\examples{
set.seed(2025)
A1 <- matrix(rnorm(10*3), 10)
A2 <- matrix(rnorm(12*3), 12)
sv <- svd(A1 \%*\% t(A2))$d
( egg <- get_gr_rank(sv) )

}
