% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini-paras.R
\name{ini_paras}
\alias{ini_paras}
\title{Initialize Parameters}
\usage{
ini_paras(
  x,
  y,
  family,
  maxRank = min(floor(NROW(x)/2), floor(NCOL(x)/2)),
  tol = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{x}{Numeric array.}

\item{y}{Numeric vector.}

\item{family}{See \link{glm} and \link{family}.}

\item{maxRank}{The maximum of rank to be detected.}

\item{tol}{Convergence tolerance.}

\item{verbose}{Print iterations?}
}
\value{
\code{list( family,  grad,  lipschitz,  sigma0,  rankStar,  rankStarHat,  lambda.star,  L0.star )}
}
\description{
Initialize \eqn{\widehat{L}} via nuclear-norm
regularized regression.
}
