% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrrglm.R
\name{nrr}
\alias{nrr}
\alias{nrrglm}
\title{Nuclear-Norm Regularized Matrix-Variate GLMs}
\usage{
nrrglm(
  x,
  y,
  family,
  lambda,
  L0 = NULL,
  tol = 0.001,
  maxIter = 500,
  verbose = FALSE
)
}
\arguments{
\item{x}{\eqn{p_1 \times p_2 \times N} numeric array with \strong{mean zero}.}

\item{y}{Numeric vector of length \eqn{N}.}

\item{family}{See \link{glm} and \link{family}.}

\item{lambda}{\eqn{\lambda}.}

\item{L0}{Initialization to \eqn{L}.}

\item{tol}{Convergence tolerance.}

\item{maxIter}{Maximal step of iterations.}

\item{verbose}{Print iterations?}
}
\value{
\code{list(L, Lsv, intercept, iter, isConvergent, lambda, deltaval)}.
}
\description{
Solve the following problem
\deqn{
 \min_{L} \Big\lbrace 
     \frac{1}{2N} \sum_{n=1}^N \ell(y_n, X_n) + \lambda \|L\|_* 
 \Big\rbrace.
}
}
\examples{
set.seed(2025)
r0 <- 13
c0 <- 17
N <- 500
x <- array(runif(r0*c0*N), c(r0, c0, N))
y <- rnorm(N)

family <- "gaussian"
lambda <- 0.15

system.time( result <- nrrglm(x, y, family, lambda, verbose=TRUE) )

}
