% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu-factor-model.R
\name{simu_factor_model_paras}
\alias{simu_factor_model_paras}
\alias{simu_factor_model_data}
\alias{simu-factor-model}
\title{Simulation: Factor Model with Correlated Noise}
\usage{
simu_factor_model_paras(p, r, noise, seed = 2023)

simu_factor_model_data(N, B, S)
}
\arguments{
\item{p}{Dimension of data.}

\item{r}{Number of factors.}

\item{noise}{Mode of \eqn{S}:
\itemize{
\item \code{diag}: Diagonal matrix.
\item \code{rand}: Nonzero elements at random positions.
\item \code{tridiag}: Tri-diagonal matrix.
\item \code{block}: Block matrix.
}}

\item{seed}{Random seed.}

\item{N}{Sample size.}

\item{B}{Matrix \eqn{p \times r}.}

\item{S}{Symmetric matrix \eqn{p \times p}.}
}
\value{
\itemize{
\item \code{simu_factor_model_paras}: \code{list(B, S)}.
\item \code{simu_factor_model_data}: Numeric matrix \code{y} of \eqn{N \times p}.
}
}
\description{
\eqn{y = B u + w} with \eqn{u \sim N(0, I_{r})},
\eqn{w \sim N(0, S)} and \eqn{B \in \mathbb{R}^{p \times r}}.

\itemize{
\item \code{simu_factor_model_paras}: simulate \verb{(B, S)}.
\item \code{simu_factor_model_data}: simulate \code{y}.
}
}
\examples{
set.seed(2025)
N <- 50
p <- 15
r <- 7

paras <- simu_factor_model_paras(p, r, 'diag')
paras <- simu_factor_model_paras(p, r, 'rand')
paras <- simu_factor_model_paras(p, r, 'tridiag')
paras <- simu_factor_model_paras(p, r, 'block')

DT <- simu_factor_model_data(N, paras[["B"]], paras[["S"]])

}
