% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu-zhouandli2014.R
\name{simu_zhouandli2014}
\alias{simu_zhouandli2014}
\title{Simulation: Setups in Regularized Matrix Regression}
\usage{
simu_zhouandli2014(N, p1, p2, r0, s0, family)
}
\arguments{
\item{N}{Sample size.}

\item{p1}{Row dimension.}

\item{p2}{Column dimension.}

\item{r0}{Rank.}

\item{s0}{signal proportion.}

\item{family}{\code{"gaussian"} or \code{"binomial"}.}
}
\value{
\code{list( C0, train=list(x, y), test=list(x, y) )}.
}
\description{
Simulation: Setups in Regularized Matrix Regression
}
\examples{
N <- 100
p1 <- 64
p2 <- 64
r0 <- 1
s0 <- 0.05
family <- 'gaussian'

set.seed(2026)
DT <- simu_zhouandli2014(N, p1, p2, r0, s0, family)

}
\references{
Zhou Hua and Li Lexin (2014). Regularized Matrix Regression. Journal of the Royal Statistical Society (Series B). \url{doi:10.1111/rssb.12031}
}
