% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drr-factor-model.R
\name{tune_drr_factor_model}
\alias{tune_drr_factor_model}
\title{Simulating Data for Factor Model with Correlated Noise}
\usage{
tune_drr_factor_model(
  y,
  lambda1.factor = 1.1^(-2:2),
  S.diag.penalize = FALSE,
  tol = 0.001,
  maxIter = 500
)
}
\arguments{
\item{y}{Numeric matrix of \eqn{N \times p}.}

\item{lambda1.factor}{Factor on \eqn{\lambda_1^*} (by default \code{1.1^(-2:2)}).}

\item{S.diag.penalize}{Whether to penalize the diagonal elements of \code{S} (by default \code{FALSE}).}

\item{tol}{Convergence tolerance.}

\item{maxIter}{Maximal step of iterations.}
}
\value{
List.
}
\description{
Doubly regularized decomposition for covariance matrix in factor model.
}
\examples{
set.seed(2025)
N <- 500
p <- 30
r <- 10
noise <- "diag"
paras <- simu_factor_model_paras(p, r, noise)
y <- simu_factor_model_data(N, paras[["B"]], paras[["S"]])
lambda1.factor <- 1.1^(-1:1)
system.time( egg <- tune_drr_factor_model(y, lambda1.factor) )

}
