% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drrglm-package.R
\docType{package}
\name{drrglm-package}
\alias{drrglm-package}
\title{drrglm: Doubly Regularized Matrix-Variate Regression}
\description{
The doubly regularized matrix-variate regression solves a low-rank-plus-sparse structure for matrix-variate generalized linear models through a weighted combination of nuclear-norm and L1-norm. The methodology implemented by this package is described in the paper "Doubly Regularized Matrix-Variate Regression", which has been tentatively accepted for publication but does not yet have a DOI or URL. A formal citation will be added in a future update once the final publication details are available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paradoxical-rhapsody/drrglm}
  \item Report bugs at \url{https://github.com/paradoxical-rhapsody/drrglm/issues}
}

}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com} [copyright holder]

Authors:
\itemize{
  \item Shan Luo
  \item Binyan Jiang
}

}
\keyword{internal}
