% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drrglm-package.R
\docType{package}
\name{drrglm-package}
\alias{drrglm-package}
\title{drrglm: Doubly Regularized Regression for Matrix-Variate Generalized Linear Models}
\description{
The doubly regularized regression for matrix-variate data solves a low-rank-plus-sparse structure for matrix-variate generalized linear models through a weighted penalty of nuclear-norm and L1-norm. See the paper by Xu, Luo and Jiang (2025, \doi{10.tbd}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paradoxical-rhapsody/drrglm}
  \item Report bugs at \url{https://github.com/paradoxical-rhapsody/drrglm/issues}
}

}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

Authors:
\itemize{
  \item Shan Luo
  \item Binyan Jiang
}

}
\keyword{internal}
