% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drrglm-package.R
\docType{package}
\name{drrglm-package}
\alias{drrglm-package}
\title{drrglm: Doubly Regularized Matrix-Variate Regression}
\description{
The doubly regularized matrix-variate regression solves a low-rank-plus-sparse structure for matrix-variate generalized linear models through a weighted combination of nuclear-norm and L1-norm. The method is detailed in the manuscript "Doubly Regularized Matrix-Variate Regression", which has been tentatively accepted by the Journal of Computational and Graphical Statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paradoxical-rhapsody/drrglm}
  \item Report bugs at \url{https://github.com/paradoxical-rhapsody/drrglm/issues}
}

}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com} [copyright holder]

Authors:
\itemize{
  \item Shan Luo
  \item Binyan Jiang
}

}
\keyword{internal}
