% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu-reg-model.R
\name{simu_reg_coefs}
\alias{simu_reg_coefs}
\alias{simu_reg_data}
\alias{simu-reg-model}
\title{Simulation: Matrix-Variate GLM}
\usage{
simu_reg_coefs(p1, p2, rank0, S.type, seed = 2023)

simu_reg_data(family, N, C0, err.student.dof = NULL)
}
\arguments{
\item{p1}{Row dimension.}

\item{p2}{Column dimension.}

\item{rank0}{Rank of \code{L}.}

\item{S.type}{Type of \code{S}.
\itemize{
\item \code{zero}: Zero matrix.
\item \code{rand}: Nonzero elements at random positions.
\item \code{block}: Block matrix.
\item \code{diag}: Diagonal matrix.
}}

\item{seed}{Random seed.}

\item{family}{See \link{glm} and \link{family}.}

\item{N}{Sample size.}

\item{C0}{Coefficient matrix.}

\item{err.student.dof}{The degree of freedom of Student-\eqn{t} for error term in
\code{family="gaussian"}. By default it is \code{NULL} for normal error.}
}
\value{
\itemize{
\item \code{simu_reg_coefs}: \code{list(L, S)}.
\item \code{simu_reg_data}: \code{list( train=list(x, y), test=list(x, y) )}.
}
}
\description{
Simulation: Matrix-Variate GLM
}
\examples{
p1 <- 10
p2 <- 9
rank0 <- 3

coefs <- simu_reg_coefs(p1, p2, rank0, "zero")
coefs <- simu_reg_coefs(p1, p2, rank0, "rand")
coefs <- simu_reg_coefs(p1, p2, rank0, "block")
coefs <- simu_reg_coefs(p1, p2, rank0, "diag")

N <- 100
S.type <- "rand"
family <- "binomial"

coefs <- simu_reg_coefs(p1, p2, rank0, S.type)
C0 <- coefs[["L"]] + coefs[["S"]]

set.seed(2025)
DT <- simu_reg_data(family, N, C0)
DT <- simu_reg_data("gaussian", N, C0, err.student.dof=3)

}
