% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-drrglm.R
\name{tune.drrglm}
\alias{tune.drrglm}
\alias{tune_drrglm}
\title{Doubly Regularized Regression for Matrix-Variate Data}
\usage{
tune_drrglm(
  iniParas,
  x,
  y,
  lambda1.factor = 1.3^seq(-3, 3, 0.2),
  maxCard = 30,
  tol = 0.001,
  maxIter = 500
)
}
\arguments{
\item{iniParas}{Initial values returned by \link{ini_paras}. See examples below.}

\item{x}{Numeric array.}

\item{y}{Numeric vector.}

\item{lambda1.factor}{Factor on initial guess\eqn{\lambda_1^2}
(By default \code{1.3^seq(-3, 3, 0.2)}).}

\item{maxCard}{The maximal cardinality of \eqn{S}.}

\item{tol}{Convergence tolerance.}

\item{maxIter}{Maximal step of iterations.}
}
\value{
List.
}
\description{
Tune the \eqn{(\lambda_1, \lambda_2)} in \link{drr}.
}
\examples{
p1 <- 10
p2 <- 9
rank0 <- 3
S.type <- "rand"
coefs <- simu_reg_coefs(p1, p2, rank0, S.type)

N <- 500
family <- "gaussian"
L0 <- coefs[["L"]]
S0 <- coefs[["S"]]
C0 <- L0 + S0
DT <- simu_reg_data(family, N, C0)

x <- DT[["train"]][["x"]]
y <- DT[["train"]][["y"]]

lambda1.factor <- lambda.factor <- 1.1^(0:1)
maxRank <- 5

system.time( iniParas <- ini_paras(x, y, family, maxRank) )
system.time( drrObj <- tune_drrglm(iniParas, x, y, lambda1.factor) )

}
